Add dropdown controllers
This patch adds two "dropdown" controllers, one for String objects, and
one for Item instances.

The controller has a list of valid values. For the item controller, this
is any registered item identifier (whether from vanilla or mods).
When the controller's string input field is selected, a dropdown is
rendered containing all options of this list that match the current
value of the input field. The list can be navigated using arrow keys.

On closing the input field (either through the "Enter" key or selecting a
different option), the input field's value is set to whichever valid
value was selected in the dropdown list.

The item controller has the bonus feature of rendering the item inside
of the option. :)

Further, to allow the item controller to work, the built-in GSON config
instance is taught to (de)serialize items to String values based on their
identifier.