Saturation/Light Picking + Bug fixes + Better mini preview outline
FEAT: Made it so that you can now choose the saturation and light values of a color! Holding down the mouse button allows you to leave the dedicated box for easier color picking. Doing the same with the hue slider doesn't affect the saturation/light picking either
REFACTOR: If a color is very bright, the mini color preview outline will change to a light grey upon hovering instead of plain white. This is to indicate to new users who may have a very bright color as an option that it can still be clicked on.
FIX: Fixed a bug where choosing a color that was too dark would result in the saturation being reset
FIX: Fixed a bug where having too dark of a color would mess up the hue slider
TEST: Tried making the background of the color picker a texture. I was unsuccessful this time, but I'm going to try again soon.