Typing bug fix and temporary float color picker fix
FIX: Fixed a bug where typing in the color controller while the color picker is visible would result in the color picker desyncing
TEMPFIX: Added a temporary workaround to the color picker floating when the color controller wasn't visible anymore. A better fix will be added in the future. The current workaround probably isn't great for performance.