Propagate right click mouse events to widgets contained in option lists (#180)
* Propagate right click mouse events to widgets contained in option lists

Starting with Minecraft 1.20.2, only left click events are passed to
list elements. This adds the required override to the respective
versions to get right click events.

With this change, right-clicking a cycling controller (like enum
controller) element properly walks backwards through the list.

Closes #157

* Propagate right-click drag events to controllers

This requires a bit of plumbing in several places, because Minecraft
aggressively filters for left-click drags.